Fix notes ahead of current note being missed first with dynamic window
This could happen if the window shrinks extremely fast and the back end miss time for notes ahead of the current note is before the back end of the current note because of the window size changing. Only the first note in the hit window can be missed on each execution.