Alternate Take On A Replay Refactor (#208)
* Refactor Replay file handling

Completely overhauls the replay file format and how the app serializes it.

Firstly, one of the described purposes on the replay versioning system was that old replays, even if out of date, to some degree, should stay legible. That is, to say, that minimally, we should be able to decipher just enough information that we can add the replay as an entry to the in-game history or gamerecords menus. This was a bit problematic beforehand as the types that handled these contained shared conflicting information. You had to first attempt to read the *entire* file - which might not even work because the old lack of proper version control -, then extract just that metadata info out to create the entry type.

With this update, I've completely separated the "metadata" information into a different structure from the actual data controlling the replay. This means that the metadata can exist independently from the replay data that it maps to, which grants a bit more in regards how we decide to deserialize replay files in certain situations.

Replay deserialization thereby gets a new "mode": metadata only. When cataloging a collection of replays, the actual replay data should only be used for simple validation against the hash. Therefore, a separate function is provided that does just that.

Secondly, the replay file has seen a major update:
1. The starting header tag has changed from YARGPLAY to YAREPLAY. This allows us to easily distinguish old replays from new ones right out the gate, allowing pre-version 6 replays to immediately communicate their "metadata only" state.
2. The replay metadata and main data now have two separate hashes. The hash for the metadata is provided at the start of the file (alongside the length of the metadata chunk) while the hash for the main replay data rests *within* the metadata chunk.
3. Additionally, the engine & replay version values have been moved into the metadata chunk as well. This adds another step of security as the hashing of the metadata will therefore disallow any changes to those values after the fact.
4. Added a header to tag to every ReplayFrame chunk. It provides an extra simple mechanism for validating data integrity.

Thirdly, the serialization and deserialization of the actual replay data has taken a more self-sufficient approach. Each structure, when appropriate, will take turns constructing themselves from the stream of the file - and vice verse with serializing *to* a file. The reason for this is simple: most of the data that we write and read to and from a file is supposed to maintain a `readonly` state. However, the state of the prior handling of the data did not make that intention clear as explicitly marking the values as `readonly` was not possible. Moving to this more-object based approach now allows that, making the logical flow of the entire process easier to grasp.

+ Additional note:
++ Removed engine logger types entirely (as per Riley's request)
++ Added functions to StreamExtensions (some from BinaryReaderWriterExtensions)
++ REQUIRES MAIN REPO UPDATE

* Place replay version constants in a singular tuple

* Some ReplayIO comments

* Add missing proKeys ReplayFrame construction

* Move PresetContainer variables to ReplayData

The overhead and maintenance of a separate type that just wraps two dictionaries isn't necessary.

+ Necessitates `ReplayFrame` to have its own `Serialize()` method - which was better in hindsight anyways.

* Move ReplayPlayerInfo variables into ReplayFrame

Remove the middleman

* Remove player & color profile IDs from ReplayFrame

* Rename replay version constants

* Just... ignore the rest of the headerstream after validation in TryLoadData()

Bruh, I'm dumb

* Add ReplayStats types for the various engines

Placed in the metatdata block to allow menus to have direct access to that information without reading the rest of the file.

* Add BaseStats method to convert to ReplayStats

* Fix incorrect ReplayLength read

* Some empty line removals

* Add ProKeys entry for ReplayAnalyzer

* Slight alteration to the maxTime selection syntax

* Add SongSpeed to ReplayInfo struct