Make TimeInStarPower rounded to ticks (#266)
The absolute value does not matter as it is calculated per tick anyway.
Not doing this could lead to TimeInStarPower not being always consistent.
We could lend on a frame within same tick with microsecond difference,
it will be processed identically, same score, same tick in SP amount
but `TimeInStarPower` would be different as it's stored as fully time dependent value anyway